Replaced PositionIntHasher with PositionLongHasher everywhere.

+++ Assets/Scripts/SimpleChunks/DataGeneration/ChunkDataGenerator.cs
@@ -8,7 +8,7 @@ namespace SimpleChunks.DataGeneration
 {
     public abstract class ChunkDataGenerator : MonoBehaviour, IChunkDataProvider
     {
-        public abstract Task<NativeParallelHashMap<int, UnsafeNativeArray<VoxelData>>> GenerateChunksRawData(
+        public abstract Task<NativeParallelHashMap<long, UnsafeNativeArray<VoxelData>>> GenerateChunksRawData(
             NativeArray<Vector3Int> generatingChunksLocalPositions,
             CancellationToken? cancellationToken = null);
     }
