+++ Assets/Scripts/SimpleChunks/DataGeneration/IChunkDataProvider.cs
@@ -10,7 +10,7 @@ namespace SimpleChunks.DataGeneration
 {
     public interface IChunkDataProvider
     {
-        public Task<NativeParallelHashMap<int, UnsafeNativeArray<VoxelData>>> GenerateChunksRawData(
+        public Task<NativeParallelHashMap<long, UnsafeNativeArray<VoxelData>>> GenerateChunksRawData(
             NativeArray<Vector3Int> generatingChunksLocalPositions, 
             CancellationToken? cancellationToken = null);
     }
